# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorData'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  usecase.SubmitFeedbackInput:
    properties:
      app_id:
        type: integer
      kategori:
        description: kendala, saran, data
        type: string
      pesan:
        type: string
    required:
    - pesan
    type: object
...
  usecase.UpdateFeedbackStatusInput:
    properties:
      catatan_admin:
        type: string
      status:
        description: pending, in_progress, resolved, closed
        type: string
    required:
    - status
    type: object
host: localhost:8080
info:
  contact:
    email: bps1200@bps.go.id
    name: Tim Inovasi BPS Provinsi Sumatera Utara
    url: https://sumut.bps.go.id
  description: Backend RESTful API untuk Portal Aplikasi Terintegrasi BPS Provinsi
    Sumatera Utara.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sumut.bps.go.id
  title: Portal Aplikasi BPS Provinsi Sumatera Utara (ALUSI) API
  version: 1.2.0
paths:
...
  /admin/feedbacks:
    get:
      description: Mengambil seluruh tiket laporan kendala, saran, dan pertanyaan
        pengguna (Admin)
      parameters:
      - description: Filter status tiket (pending, in_progress, resolved, closed)
        in: query
        name: status
        type: string
      - description: Filter ID aplikasi
        in: query
        name: app_id
        type: integer
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah per halaman (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Daftar laporan feedback (Admin)
      tags:
      - Admin - Feedback
  /admin/feedbacks/{id}:
    put:
      consumes:
      - application/json
      description: Mengubah status dan menambahkan catatan respon tindak lanjut pada
        tiket laporan (Admin)
      parameters:
      - description: ID Feedback
        in: path
        name: id
        required: true
        type: integer
      - description: Data status dan respon tanggapan
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateFeedbackStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Perbarui status tindak lanjut tiket feedback (Admin)
      tags:
      - Admin - Feedback
...
  /apps:
    get:
      description: Mengambil daftar aplikasi dengan filter kategori, target pengguna,
        status, persona, dan pagination
      parameters:
      - description: Filter slug kategori
        in: query
        name: category
        type: string
      - description: Filter target pengguna (semua, internal, internal_bps, mitra,
          pimpinan)
        in: query
        name: target_pengguna
        type: string
      - description: Filter status operasional (operasional, pemeliharaan, kendala)
        in: query
        name: status
        type: string
      - description: Filter persona / tag
        in: query
        name: persona
        type: string
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah item per halaman (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Daftar katalog aplikasi
      tags:
      - Catalog
...
  /apps/{slug}:
    get:
      description: Mengambil metadata lengkap, deskripsi, panduan, dan status sebuah
        aplikasi
      parameters:
      - description: Slug aplikasi
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Detail aplikasi berdasarkan slug
      tags:
      - Catalog
  /apps/{slug}/guides:
    get:
      description: Mengambil daftar petunjuk teknis / user manual / FAQ untuk aplikasi
        terkait
      parameters:
      - description: Slug aplikasi
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Panduan dan FAQ aplikasi
      tags:
      - Catalog
...
  /apps/search:
    get:
      description: Pencarian aplikasi dengan toleransi typo, sinonim kata kunci, dan
        filter kategori
      parameters:
      - description: Kata kunci pencarian
        in: query
        name: q
        required: true
        type: string
      - description: Filter slug kategori
        in: query
        name: category
        type: string
      - description: Batas jumlah hasil (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Pencarian cerdas aplikasi (Trigram pg_trgm)
      tags:
      - Catalog
...
  /categories:
    get:
      description: Mengambil seluruh kategori aplikasi yang aktif
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Daftar kategori aplikasi
      tags:
      - Catalog
  /categories/{slug}:
    get:
      description: Mengambil detail data sebuah kategori aplikasi
      parameters:
      - description: Slug kategori
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Detail kategori berdasarkan slug
      tags:
      - Catalog
  /feedbacks:
    post:
      consumes:
      - application/json
      description: Mengirim formulir laporan kendala atau usulan fitur untuk aplikasi
        tertentu atau portal secara umum
      parameters:
      - description: Data formulir feedback
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.SubmitFeedbackInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Kirim kritik, saran, atau laporan kendala
      tags:
      - Feedback
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"